Use the current hostname when accessing the API
The Content Security Policy rules combined with the fact that the
API is currently on a separate port makes using the web interface
from the Pi-hole's IP address difficult. I have yet to find a way
to allow the `'self'` origin on all ports without hardcoding in
the Pi-hole's IP address. The CSP spec recommends to use hostnames
rather than IP addresses, so unless the API is reverse-proxied or
we modify the CSP metadata whenever the IP address is changed,
users will not be able to use the web interface via IP address.